refactor: project structure
added /internal, /pkg and /config to follow go best practices. Created models package and moved parser, validation and transform to /pkg
