+++ config/config.go
@@ -1,0 +1,35 @@
+package config
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"time"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/mongo/readpref"
+)
+
+var mongoClient *mongo.Client
+
+func ConnectDb() (*mongo.Client, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	mongoUri := os.Getenv("MONGO_URI")
+	clientOptions := options.Client().SetTimeout(3 * time.Second).ApplyURI(mongoUri)
+
+	client, err := mongo.Connect(ctx, clientOptions)
+	if err != nil {
+		return nil, fmt.Errorf("error while connecting to mongoDB: %v", err)
+	}
+
+	err = client.Ping(ctx, readpref.Primary())
+	if err != nil {
+		return nil, fmt.Errorf("error while pinging the database: %v", err)
+	}
+
+	fmt.Println("Connected succesfuly to mongoDB!")
+	return client, nil
+}
