+++ config/loadenv.go
@@ -1,0 +1,7 @@
+package config
+
+import "os"
+
+func GetMongoUri() string {
+	return os.Getenv("MONGO_URI")
+}
