+++ pkg/transform/transformer.go
@@ -1,25 +1,25 @@
-package transformer
+package transform
 
 import (
 	"strings"
 
-	"github.com/MarcinZ20/bankAPI/handlers/parser"
+	"github.com/MarcinZ20/bankAPI/pkg/models"
 )
 
 // IsBranchOf checks if a bank is a branch of a headquarters
-func IsBranchOf(bank parser.Bank, hq Headquarter) bool {
+func IsBranchOf(bank models.Bank, hq models.Headquarter) bool {
 	return bank.SwiftCode[:8] == hq.SwiftCode[:8] && bank.CountryISO2Code == hq.CountryISO2Code
 }
 
-func TransformBankEntity(bank *parser.Bank) {
+func TransformBankEntity(bank *models.Bank) {
 	bank.SwiftCode = strings.ToUpper(bank.SwiftCode)
 	bank.CountryISO2Code = strings.ToUpper(bank.CountryISO2Code)
 	bank.Name = strings.ToUpper(bank.Name)
 	bank.CountryName = strings.ToUpper(bank.CountryName)
 }
 
-func TransformIntoHeadquarter(bank parser.Bank) Headquarter {
-	return Headquarter{
+func TransformIntoHeadquarter(bank models.Bank) models.Headquarter {
+	return models.Headquarter{
 		Address:         bank.Address,
 		BankName:        bank.Name,
 		CountryISO2Code: bank.CountryISO2Code,
@@ -29,8 +29,8 @@ func TransformIntoHeadquarter(bank parser.Bank) Headquarter {
 	}
 }
 
-func TransformIntoBranch(bank parser.Bank) Branch {
-	return Branch{
+func TransformIntoBranch(bank models.Bank) models.Branch {
+	return models.Branch{
 		Address:         bank.Address,
 		Name:            bank.Name,
 		CountryISO2Code: bank.CountryISO2Code,
@@ -40,9 +40,9 @@ func TransformIntoBranch(bank parser.Bank) Branch {
 	}
 }
 
-func Transform(banks []parser.Bank) map[string]Headquarter {
-	hqs := make(map[string]Headquarter)
-	brs := make(map[string][]Branch)
+func Transform(banks []models.Bank) map[string]models.Headquarter {
+	hqs := make(map[string]models.Headquarter)
+	brs := make(map[string][]models.Branch)
 
 	ConvertIntoMaps(banks, hqs, brs)
 	MergeMaps(hqs, brs)
@@ -50,7 +50,7 @@ func Transform(banks []parser.Bank) map[string]Headquarter {
 	return hqs
 }
 
-func ConvertIntoMaps(banks []parser.Bank, hqs map[string]Headquarter, brs map[string][]Branch) {
+func ConvertIntoMaps(banks []models.Bank, hqs map[string]models.Headquarter, brs map[string][]models.Branch) {
 	for _, bank := range banks {
 		key_code := bank.SwiftCode[0:8]
 
@@ -63,7 +63,7 @@ func ConvertIntoMaps(banks []parser.Bank, hqs map[string]Headquarter, brs map[st
 }
 
 // TODO: Refactor using pointers
-func MergeMaps(hqs map[string]Headquarter, brs map[string][]Branch) {
+func MergeMaps(hqs map[string]models.Headquarter, brs map[string][]models.Branch) {
 	for key_code := range hqs {
 		if branches, ok := brs[key_code]; ok {
 			hq := hqs[key_code]
